migrate: Ensure lookup obligations are satisfied for remapped nodes
When we create duplicate nodes to prevent fully materialized nodes from
being added as children of partially materialized nodes (as introduced
in 8d48e9add (Fix Partial Materialization Bug, 2021-10-13)), we need to
make sure that those newly created nodes get (full) indices added to
them to support any lookup obligations in their existing children. The
easiest way to make this happen is to just add the child to the set of
`new_nodes` so that it gets looked at in the next iteration of the
migration loop. The one caveat here is that this also means we need to
do some extra idempotency checking to make sure we don't also generate a
new local index for that node (which is probably a good idea anyway,
since idempotency is generally a good thing).

This also modifies the existing `reroutes_recursively` integration test
to trigger the original cause of this bug.

Fixes: ENG-1713
Change-Id: I00794c296051f8a62c271767c9ee640ad725a919
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3488
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>