Add `DfType::DEFAULT_TEXT` constant
This avoids needing to specify `Collation::default()`, which makes
related code easier to read.

This has the added performance benefit that when passed directly by
reference (e.g. `DfValue::coerce_to`), the compiler is allowed to use a
constant defined in the binary rather than on the stack. This was not
possible before because `Default::default()` cannot currently be
const-evaluated.

Change-Id: Ib3cad2f281d3be10660287dbd1d231265a0a533b
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3492
Reviewed-by: Griffin Smith <griffin@readyset.io>
Tested-by: Buildkite CI