psql-srv: Fix querying upstream for type information
`load_extended_types` in psql-srv, which is called to query information
for unsupported custom types from the upstream, was implemented using a
simple query but only pattern matched on UpstreamResponse::Select,
panicking with a `todo!()` (!!) on any other variant - including the
UpstreamResponse::SimpleQuery that was actually getting returned. This
fixes that to only handle SimpleQuery, and covers proxying unsupported
custom types with a test.

Change-Id: Ia373d5a20f9a1e2f4fbba6c65cd711dde9dc134a
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3499
Reviewed-by: Dan Wilbanks <dan@readyset.io>
Tested-by: Buildkite CI