readyset-data: Remove unused DfType::dialect method
There are certainly places in the code where we match on the dialect of
a specific type variant, but this function is currently unused and there
do not seem to be any places where we want to know the dialect of a type
irrespective of the specific DfType variant. I don't think this is
likely to change in the future, so I am removing the unused code here.

Change-Id: Ifb9388b29be37c8686b38682b1bd3d1bd03cc121
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3501
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>