readyset-data: Remove dialect field from DfType::Json variant
This is not currently used for anything, so it can be easily removed.
For future code that deals with differences in MySQL and Postgres JSON
behavior, I believe it makes more sense to create different
versions of the JSON operators/functions at lowering time; the storage
semantics of the type do not change, so I don't think there's any need
to track a dialect at the type level in this way.

Change-Id: If1bd9cd653105a84f6cd1cd439badb4983d71fc1
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3502
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>