readyset-psql: Add a test for enums
Add a test to the readyset-psql types integration suite covering enums
both as parameters and returned from queries, and covering the sorting
of enum values according to postgres's semantics. Currently, since enums
are unsupported, these queries all proxy upstream, but this test serves
both as a check to make sure that as we add support for postgres enums
we maintain the same behavior as postgres itself, and as a starting
point to check that queries with postgres enums are in fact cacheable.

Change-Id: I4c7d441f84c179d17acdc53d0cf2bbde0a74c8e9
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3505
Tested-by: Buildkite CI
Reviewed-by: Nick Marino <nick@readyset.io>