replicators: Provide psql table metadata as structured data
Currently, to replicate CREATE TABLE events in postgresql, we make a
vague attempt to format the events as *mysql-compatible* (!!) strings,
then *parse* those strings to send the actual AST as part of the
ChangeList to the controller. This has always been a bit of a hack (not
to mention that it's a little spooky, as something might stop working in
the round-trip at any point in the future), but actually totally breaks
down in the case of custom types (enums, in this case) since those
actually *don't parse at all* in the MySQL dialect. Rather than making
them parse in the MySQL dialect, I figured now's as good a time as any
to get rid of the stringification and re-parsing for Postgres (not for
MySQL, since MySQL DDL events actually come in natively as strings via
the binlog) - now, instead, the DDL replication log sends metadata about
created tables as *structured data* (JSON, specifically) and we convert
that directly to the AST `CreateTableStatement`, without having to go
through the parser at all. This also requires changing the `DdlChange`
replication action to natively contain a `Vec<Change>` instead of
containing a DDL string - for MySQL this means pushing the parsing into
the `mysql_connector` instead of the `noria_adapter`

Refs: ENG-1991
Change-Id: I73c4ea899e4c2bf27a720dac1d1d79f3d1940fb3
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3510
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>