replicators: Dont snapshot tbls with gen cols
We currently dont support tables with generated columns--an insert into
such a table may only have n columns, but the table itself could have m
> n columns, and we return a WrongColumnCount error if we try to
> replicate, for example, an insert where the columns don't match.

Given this, this change extends the query for tables we should snapshot
to filter out any tables that have generated columns, so we ignore them.

An integration test was added to make sure we properly handle these
tables.

Change-Id: Ieadb23edabf9658ce9e4d4440350afc4152be986
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3512
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>