dataflow-expression: Break out a LowerContext trait
In preparation for adding a second kind of thing that can be resolved
during expression lowering (custom types), refactor the `resolve_column`
function passed into `DfExpr::lower` to a trait which implements a
method with the same name and signature. This scales slightly nicer than
just passing a bunch of different functions to `Expr::lower`.

Refs: ENG-2000
Change-Id: I626796d68250c86e2a04daf88ec169d9bfa9cab8
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3518
Tested-by: Buildkite CI
Reviewed-by: Nick Marino <nick@readyset.io>