nom-sql: Allow custom types to be schema-qualified
Custom types in postgresql live inside of a schema, and two types can
exist with the same name but in a different schema. To support that,
this commit modifies SqlType::Other and DfType::PassThrough to contain a
`Relation` instead of just a bare SqlIdentifier, threading it through
everywhere appropriate (including setting `schema` during replication).

Change-Id: I482a3f05c908a5dd4ea74c32902c1b659019b446
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3519
Tested-by: Buildkite CI
Reviewed-by: Sif Hall <sif@readyset.io>
Reviewed-by: Nick Marino <nick@readyset.io>