dataflow-expression: Try to resolve custom types during lowering
Attempt to resolve custom types in dataflow-expression during lowering,
via a new method on the LowerContext trait which can return a DfType
given a Relation, and a new function passed to `DfType::from_sql_type`.
To allow unsupported (or unknown) custom types to still work via the
passthrough mechanism, this resolution is fault-tolerant - if we don't
find a custom type with a particular name, we just fall back to
DfType::PassThrough like we did before.

Everywhere this resolution function is actually defined just always
returns None as of this commit - for the parts of those that're hit on
an actual production-facing code path this'll later be extended to look
up the custom type in some sort of registry, but many of these are only
used in testing-only code paths: specifically, the parts of the adapter
that handle inserts and updates directly when an upstream is absent.
Just saying we don't support custom types for these code paths is fine
for now, but going forward I'd like to refactor this such that the type
resolution (and coercing, which is what the type resolution is for)
happens in the server rather than the adapter.

Fixes: ENG-2000
Change-Id: I521307e192fe4fe4a8962c7464a6772ce9e59e85
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3525
Tested-by: Buildkite CI
Reviewed-by: Nick Marino <nick@readyset.io>