server: Store a map of custom types in SqlIncorporator
Add a new `custom_types` HashMap to SqlIncorporator, which
maps (schema-qualified) names for types to the actual definitions of
those types (represented as just a DfType, since that's what we'll
actually need later when doing lookups into this map), and a new variant
to `recipe::Change` to create a new named custom type. Currently nothing
uses this map, and nothing actually sends this Change, but that'll be
changing (ha) soon.

Fixes: ENG-1998
Change-Id: I2af3061d14681c6b3272eb1dc32370f32fc3012b
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3532
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>