Add disk modeling
This commit adds disk modeling as an option for the fallback cache so we
can test how the fallback cache would perform if we ran it off disk. It
accomplishes this by taking the average read/write speed for spinning
disk, which is in mb/s, and uses that to create an async sleep when
interacting with the cache, based on the caches in memory size to
roughly model average read/write speeds to/from spinning disk.

Change-Id: I79301e7c13201ce6da89a14f9e70ce64b747a8e2
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3536
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke.o@readyset.io>