readyset-data: Remove DfType::PassThrough
It turns out that the only place where DfType::PassThrough ends up,
which is the column schema construction in readyset-psql, isn't actually
able to *do* anything with it, since we don't have enough data to
construct a full `postgres_types::Type`. This has previously not been an
issue since we actually fail earlier on when replicating DDL (I believe
accidentally, since what's going on is that we *always* parse DDL using
the MySQL dialect, even when it's coming from postgres, and we fail to
parse custom types in the MySQL dialect) but will become an issue for
truly unsupported types once we start actually replicating custom types.
Since we don't actually have a use for DfType::PassThrough anymore, this
just removes it. This requires making DfType::from_sql_type fallible,
which ends up being pretty mechanical and straightforward to thread
through everything relevant.

Change-Id: Ibd264210a3fc9d5ccaed94be518a8ff3d60af255
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3548
Tested-by: Buildkite CI
Reviewed-by: Sif Hall <sif@readyset.io>