readyset-data: Store oid on DfType::Enum for pg enums
For DfType::Enum types which are replicated from postgresql, store the
OID of the type itself as a field on the `DfType::Enum` variant. Later,
we'll use this OID when converting enum types to result columns in
readyset-psql, so that the results are returned with the same OID in the
colum header as if they had come from the upstream. To improve clarity,
this also renames the previous `EnumType` struct to `EnumVariants`, to
make it clear that it only stores the variants of the enum and not the
full information about the type.

Fixes: ENG-2001
Change-Id: I3f33990bce6fd0e3031bd8904d096836f8c8f37c
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3550
Tested-by: Buildkite CI
Reviewed-by: Nick Marino <nick@readyset.io>
Reviewed-by: Nikolai Vazquez <nikolai@readyset.io>