eval: Add support for arrays to the `?` operator
This extends the existing code which only works for JSON objects, and
adds support for checking whether a string JSON value exists as a JSON
array element as well. This commit also adds a couple of corresponding
test cases. With the addition of array support, this operator is now
fully functional.

Fixes: ENG-1494
Change-Id: I8a2b89de3046ae4f7eeab79238db8cb9a99e1d9b
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3558
Tested-by: Buildkite CI
Reviewed-by: Nikolai Vazquez <nikolai@readyset.io>