data: Carry name and schema on enum types, in addition to oid
In order to construct the full postgres Type metadata to return enum
columns from readyset-psql, we need the name and schema of the type to
be carried around, in addition to just the oid. This adds those,
wrapping them up alongside oid in a new `PgEnumMetadata` struct which is
optionally stored on `DfType::Enum`. This makes it even more likely that
at some point in the near future we'll just remove `dialect` from the
type as well.

Refs: ENG-2012
Change-Id: Ic7dfa53ebc870c2d71d90dc3dba6b47db9b89e4b
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3565
Tested-by: Buildkite CI
Reviewed-by: Nick Marino <nick@readyset.io>