replicators: Delay conversion from oid to Type
Rather than converting from type oid to Type for ColumnSpec directly in
the WAL reader (which requires us to *make up* a type for the case
whwere the conversion fails due to the type not being a type that's
built-in to tokio-postgres), just store the oid on ColumnSpec directly,
and wait to convert the type to a Type until we're actually parsing
replicated values. This has no effect atm (besides avoiding the slight
hack of having to make up a type) but will later unlock conversion to
custom types, since those are most naturally stored via a field on the
WalReader itself.

Refs: ENG-2002
Change-Id: I2ce1683b073d7fbe89ab2254325df9aeeae77469
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3578
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>
Reviewed-by: Nick Marino <nick@readyset.io>