ci: Send BUILDKITE_COMMIT to builds
This change passes the BUILDKITE_COMMIT env var into the docker images
that build the server and adapter, so that it can be read by the
readyset-version crate. The commit ID is included in the --version
information.

This also changes the behavior of env_or_unknown_with_fallback() in
readyset-version/build.rs, which now treats an env var set to the empty
string as being unset.

Change-Id: I0d7807a5aeca8ac9b05bab0267e18a0cbbdc0a39
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3583
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke.o@readyset.io>