replicators: Send custom types as raw Text
When reading raw records from the WAL, it isn't trivial to get enough
information to fully coerce values with custom types to the target
type (eg for enums converting the enum label text to the integer enum
index). Instead, this commit just sends values of custom types as their
raw text - this works for enums since we're *already* coercing them to
the target type within the table itself. This approach ostensibly
slightly less efficient only because the string value is a bit larger
than the integer index, but I doubt it's a big deal.

To make this all work, the WalReader now records a set of the oids of
all custom types it's been told about via the Type WalRecord, passing
this into the `into_noria_vec` method.

Refs: ENG-2002
Change-Id: I241306e38958ed9577c7bab129c1dd4cacd811e3
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3579
Tested-by: Buildkite CI
Reviewed-by: Nick Marino <nick@readyset.io>