dataflow-expression: Coerce right value to left type when comparing
When evaluating Equal and NotEqual expressions, coerce the value on the
rhs to the type of the lhs before comparing, so that eg enum values
compare as equal to their corresponding string labels.

In the future, we may want to do this via adding an extra Cast node at
lowering time - and in fact we may want to do it for other operators as
well - but in the meantime this is at least *correct* until we get a
chance to more thoroughly evaluate where mysql and postgres do implicit
coercions during equality testing.

Change-Id: Ide17086f2e7431672086c271233e0497efe39e31
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3589
Tested-by: Buildkite CI
Reviewed-by: Nick Marino <nick@readyset.io>