replicators: Don't stringify and re-parse snapshotted tables
Similarly to what we're doing for *replicated* tables, rather than
sending CREATE TABLE statements that we get during snapshotting to the
controller by stringifying then re-parsing them, just construct the
Change directly (the only thing we *do* need to parse is the SqlType,
but this is pretty trivial now). This similarly avoids the issue where
custom types in tables failed to snapshot because we aren't able to
parse custom types with the MySQL dialect.

Change-Id: Ia554735f5610fe53b9c2c7660c2c006bc7b4473f
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3590
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>