Inlining: Add ExprSkeleton to Registry
Creates the ExprSkeletons struct to represent cached queries by their
structure, without regard for literal values, and adds queries to this
cache when they are migrated. The structure will be used to determine
whether a cached query can satisfy a request for a View, for some set of
parameters passed to the query for which the View was requested.

For example, if a migration exists for SELECT * FROM t WHERE x = 1 and y
= ?, we would like to know that the cached query could be used for
either 'SELECT * FROM t WHERE x = ? and y = ?' or 'SELECT * FROM t WHERE
x = ?  and y = 1', depending on the literal values passed during query
execution.

ExprSkeletons will be used to record that a cached query exists for the
query 'SELECT * FROM t WHERE x = ? and y = ?' with the literal values
(1, ?). Then, if we receive queries with the same structure but
different literal values, as in the example above, we can compare the
literal values to determine whether the cached query could be used for
the given query.

If there exist multiple cached query with the same structure, but
different literal values, we can also determine which of these cached
queries could satisfy a query for which a view is requested with yet
another set of literal values.

A MatchedCache struct is created to represent queries that could re-use
an existing migration, and the relationship between those queries'
literals and the cached query's literals.

Change-Id: Ib4750c13eb601c00ff13d7ce058a655899dc5ca7
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3415
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>
Reviewed-by: Fran Noriega <fran@readyset.io>