Inlining: Store compatible migrations during extend_recipe
When a query fails migration during an extend_recipe call, the
controller now checks to see if the query can reuse an existing cache.
This is useful when the query given to extend_recipe contains
placeholder variables in locations not supported by ReadySet, but there
exists a cache with inlined values at the location of the placeholders.

If a cache exists that could satisfy lookups for the given query, we no
longer return an error. Instead, we log a warning that the migration
failed and record the cache that can be reused for this query in a new
field in ExprRegistry.

Also updates remove_expression to cleanup Registry::skeletons and
Registry::reused_caches. Since reused caches only live in the Registry,
we do not have to clean up the graph when we receive a request to remove
a reused cache.

Change-Id: Ia0a4fe1b6ba913e9351c89eff44158298be39044
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3652
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>
Reviewed-by: Fran Noriega <fran@readyset.io>