query_graph: Fix error for certain JOIN predicates specified in WHERE
MySQL is unusual in that you are allowed to omit the ON condition when
using JOIN or INNER JOIN. This effectively results in a cross join.
However, it is then possible to specify a multi-table condition in the
WHERE clause which effectively turns the result of the cross join back
into an inner join.

Ultimately, this means that all of these queries are effectively the
same:

SELECT * FROM a JOIN b ON a.b_id = b.id;
SELECT * FROM a JOIN b WHERE a.b_id = b.id;
SELECT * FROM a, b WHERE a.b_id = b.id;

We already have code to correctly handle the third case (the "implied
join", also known as the "comma join") by looking for predicates in
`classify_conditionals` that do an equality comparison on columns from
two different tables. Such predicates then get saved in
`join_predicates` and we end up inserting them back into the `joins`
part of the query graph later on.

However, before this CL, the code in `classify_conditionals` would only
look at the `tables` field from the `SelectStatement`, which only
contains tables mentioned directly in the FROM clause, and does not
include tables (or other relations) mentioned in the JOIN clause. This
means that the second of the three example queries above would fail to
be handled correctly and resulted in an error message about an invariant
being violated in the `set_names` function in mir_to_flow.rs.

To fix this, this commit adds a new HashSet in `to_query_graph` called
`inner_join_relations` which tracks the list of relations that are
mentioned in WHERE, JOIN, and INNER JOIN clauses. This allows us to
correctly handle join predicates for the second of the three example
queries shown above, as well as similar queries that join on subqueries
instead of tables.

More detailed examples are included in the newly added logictest that is
included in this commit. The first of the three SELECT queries in that
test already work, but the latter two queries fail without the changes
that this commit makes.

Fixes: ENG-2016
Change-Id: Iec342dba3d691981f37b50973f72b41177759ff4
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3598
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>