nom-sql: Enum variants can only be String
In both mysql[0] and postgres[1] enum variants are limited to only
containing string literals - whereas our enum representation was
using a vector of any Literal value. This commit changes the
EnumVariants struct to instead wrap a list of String instead of Literal,
which simplifies some code that has to match on enum variants when
coercing and when converting to pg type metadata.

[0]: https://dev.mysql.com/doc/refman/8.0/en/enum.html#enum-using
[1]: https://www.postgresql.org/docs/current/datatype-enum.html

Change-Id: I910cdd09a23725992976bda78ec21341cf029eb2
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3600
Tested-by: Buildkite CI
Reviewed-by: Nick Marino <nick@readyset.io>