Replace use of mem::replace to take Expr values with new `take` method
There are a few places in the code where we use mem::replace to move an
Expr value into a new location, and replace the old location with
Expr::Literal(Literal::Null). This code pattern is a bit wordy and
potentially confusing in the way that we construct a new expression that
doesn't actually get used anywhere, so I've simplified these cases by
implementing a new `take` method on Expr to encapsulate this pattern.

Change-Id: I5e958aea0453077440146fce21d84f38e952a40f
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3602
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>