dataflow-expression: Rename QuestionMark op to JsonExists
Note that we are still naming the operator QuestionMark in nom_sql,
because it's possible that the same operator might mean different things
in other databases, and the parser tries to be agnostic of the
underlying semantics. But it makes the dataflow code clearer if we lower
the AST QuestionMark to an operator with a name that reflects what it
actually does, rather than being named after what it looks like.

Change-Id: Ic2dbfe85f5468c0780ebf6747485f0f6fe5e9dd3
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3610
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>
Reviewed-by: Griffin Smith <griffin@readyset.io>