server: Add Change to add new variants to a type
Add a new variant to Change, AlterType, which itself contains a name of
a type to alter and an enum for the change to make to that type.
Currently, that enum only has one variant, SetVariants, which allows
setting the variants of a type, with the (current) limitation that we
can only add new variants to the end of the enum, and can't drop
variants, reorder variants, or add new variants inside the existing list
of variants (this limitation will be lifted later, but for now this
avoids the extra complexity of having to do a bulk table update to
change existing variant indices).

Since the type information is added as metadata to all dependent nodes,
we *do* still need to drop all nodes and queries that're downstream of
any tables that contain columns with that type, which we do using a
similar method to what we do for ALTER TABLE / DROP TABLE.

Refs: ENG-2003
Change-Id: I27f9a5f86df528043e81d0ac19d1b65d23950611
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3627
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>