dataflow: Add DomainRequest to change the type of a column
Add a new DomainRequest::SetColumnType message, which will change the
type of a column in an arbitrary node within the domain. This will be
used for adding new variants to the end of existing enum types, to
change the types of the columns within base nodes that have columns of
that enum type.

Refs: ENG-2003
Change-Id: I9848f273f7068f34899ffef37bfeea3dc78b3a75
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3624
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>