recipe: Track tables with columns referencing custom types
Track a map in the expression registry from names of custom types, to
tables containing columns that either have that type, or reference that
type (eg via an array of that type). Later, this'll be used when we
ALTER that custom type to update the type on that table's columns, and
even later to update the values within those columns in-place.

Refs: ENG-2003
Change-Id: I239fbdf50d82d12f155bbce7c3218a25ece9430d
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3625
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>