Add eviction modeling to fallback cache
This commit adds eviction modeling to the fallback cache. It allows a
user to specify an eviction rate as a floating point percentage. That
eviction rate indicates how often we will evict a query. It works as
such:
1. For every lookup that happens to the fallback cache, a counter is
   incremented.
2. Whenever that counter hits an even number that falls in line with the
   eviction rate, the query is evicted from the fallback cache. For
   sample, if you have a 0.01 eviction rate (1%), then 1 in every 100
   lookups will cause an eviction to occur. with a 0.5 eviction rate
   (50%), one in every 2 lookups will cause an eviction.
This is very naive but allows us to capture data on how evictions would
impact performance.

Change-Id: Idda303e49cf70c662b8e41ee037f9abc2a8d416d
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3650
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke.o@readyset.io>