nom-sql: Split visit and visit_mut
Currently, our Visitor trait in nom-sql takes mutable references to the
AST structs that it visits - this has worked for us up to this point,
but I'm about to need to do something that requires only passing shared
references. To support that, this commit renames our existing `visit`
module to `visit_mut`, and `Visitor` to `VisitorMut`, and
then *duplicates* it (urgh) to a new `visit` module, which takes
shared (immutable) references to everything.

This duplication is a little annoying - it might be nice to do with a
macro, but that macro would likely be *very* complex - but hopefully our
AST grows rarely enough that having to add visitor methods in two
different places isn't going to be too bad. Someday keyword generics
will land and this can all be merged into one again.

Change-Id: I81e8eaed19049285d800e1feb85fc80e5d6d0c8c
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3629
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke.o@readyset.io>