recipe: Track queries that reference custom types too
Queries can reference custom types in addition to just
tables (specifically in CAST expressions, right now) - when we alter
types, we need to invalidate those queries too. This commit doesn't do
that directly, but it does pave the road to doing that by tracking
which queries reference which custom types in the ExprRegistry.

Change-Id: I43f2d1dea9578874b366c8f017c83936e0629187
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3630
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>