Update tracing deps to current versions
This `cargo updates` the various tracing packages to the current
versions. This brings up things like updated `instrument` macros
which allow better control & flexability.

In `plan.rs`, `clippy::unreachable` was allowed on `add`, due to
the `instrument` macro having a usage of `unreachable`. This has
been dealt with upstream and should be included in a future release
of tracing.

Updated packages:
 * tracing
 * tracing-subscriber
 * tracing-futures
 * tracing-opentelemetry
 * tracing-attributes

Change-Id: I477e7714227f1d2e4815dfb169fe8aafed2ef8f9
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3632
Tested-by: Buildkite CI
Reviewed-by: Peter Farr <peter@readyset.io>