test-utils: Add utility for limiting test parallelism
Add a new method to the test-utils crate for limiting the number of
(all sync or async) tests within a group that can run in parallel.
Internally, this uses a tokio::sync::Semaphore in the async case, and
a (crappy implementation of a) non-async semaphore in the sync case,
since there doesn't seem to be a good rust semaphore out there that fits
our use cases (specifically, a `const fn new` and good blocking
characteristics). The latter uses a spin-loop instead of a waiter queue
with condvars for simplicity, but if that gets too slow we should
do the waiter queue later.

This is likely generally useful, but I'm doing it right now for the
replicators test suite, since that has to limit the number of concurrent
tests that run due to postgres's limitation on the number of replication
slots that can exist at once.

Change-Id: Id9948c9d219993b6a637fbbbe87cd9d166c0c5a2
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3643
Tested-by: Buildkite CI
Reviewed-by: Nick Marino <nick@readyset.io>