replicators: Mitigate limitation on replication slots in tests
PostgreSQL has an upper limit on the number of replication slots that
can exist in a single db, which we're about to hit in the test suite for
`ddl_replication.rs` in replicators. To keep us running at least *a
little bit* in parallel, this commit uses the new `parallel_group`
utility to limit the number of these tests which we can run in parallel,
dropping the replication slot at the end of each test to allow
subsequent tests to create a slot without hitting the limit.

Change-Id: Idecf70a740b55b519bb2528b297a360144182795
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3644
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>