dataflow-expression: Implement full support for `?&` operator
This adds relevant code and tests to `eval` to implement `?&`, as well
as some necessary boilerplate code in `lower` and `normalize_negation`.

The `?&` operator is similar enough to `?|` that it was more succinct to
incorporate the new evaluation code into the existing clause for `?|`
rather than adding a separate new clause and copying/pasting the
relevant parts of the code.

Change-Id: Id6abe8cb1fdbc58e2611b33335bace2f49dffae2
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3646
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>
Reviewed-by: Griffin Smith <griffin@readyset.io>