Inlining: Add required_values and key_remapping to View
Adds two additional fields to View that allows a View to be reused when
the executed and cached query share the same structure, but different
literal values. These fields are only Some(_) if the View is reused by a
query that is different than the cached query.

required_values is a map of every placeholder in the executed query that
maps to an inlined value in the cached query. These values must be
passed to the executed query at query execution or the query will be
sent to fallback.

key_remapping is a map from placeholders in the cached query to literals
in the executed query (which may include placeholders). Every
placeholder in the cached query will appear in this map. Every
placeholder in the executed query will either appear in this map or the
required_values map.

This change also adds the logic to build a ViewQuery against a reused
View.

Change-Id: I6aa02647c4f076abbc905e9ce8c340dcd6f7f92b
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3654
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>