nom-sql: ARRAY is an expression, not a literal
The postgresql ARRAY expression can take other expressions within its
subexpressions, eg:

    localhost/postgres=# select array[1, '2'::int];
     array
    -------
     {1,2}

Given that, it shouldn't be a variant of Literal in nom-sql, it should
be a variant of Expr, and the vec of elements should be a vec of Expr,
not a vec of Literal. This commit makes that change, which involves also
adding a new Array variant to DfExpr as well - to avoid having to do any
recursing into nested vectors at eval-time, this flattens out the vector
of element expressions and provides a `shape` vec, which is what ArrayD
expects at construction time.

Refs: ENG-2050
Change-Id: Id7ad5870d55a834646c0604a1e20319dcd104d72
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3655
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>