replicators: Record the oid for arrays of custom types
In postgres, the type for "arrays of a type T" is a *distinct* type,
with a distinct row in the `pg_type` catalog table and a distinct oid.
To pave the road for support for arrays of custom types, this commit
adds a new `array_oid` field to PgEnumMetadata for the oid of *arrays of
that enum type*, and records that oid as part of replicating and
snapshotting custom types in the replicator. Currently, the field is
unused, however.

Refs: ENG-2050
Change-Id: Ib0e2d79a3ebd71acca997d90965410f5eb518c10
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3656
Tested-by: Buildkite CI
Reviewed-by: Nick Marino <nick@readyset.io>