nom-sql: Include query in err if failed to parse
If a query failed to parse, it will be useful to have a log of what the
query was, but the query may also contain sensitive information.

This change includes the query if the redact_sensitive feature is
disabled (which it is by default), otherwise it logs <redacted> in the
"failed to parse" error message.

Change-Id: I7e6639aa9e5c3af2f046fd3d55395d8e621cdbe9
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3664
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>