nom-sql: Add parsing support for `||` JSON operator
This one is slightly trickier than the other JSON parsing CLs I've
submitted recently, because `||` can represent a couple of other
operators besides JSON concat. We don't support any of those operators
yet, though, and I think it should be relatively straightforward to add
them in the future, so beyond adding some additional comments and such
this looks pretty similar to the parsing code and tests that I've added
for other JSON operators in recent weeks.

Change-Id: Id2cced29fadf5a48ce187b8dc8cab76f9280d7dc
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3670
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>
Reviewed-by: Nikolai Vazquez <nikolai@readyset.io>