Refactor: Replace get_parameter_columns_recurse() with Visitor
Replaces the get_parameter_columns_recurse() util function with a
visitor. This is a refactor and does not introduce new functionality.

This visitor returns the columns that are compared to the
placeholder parameters and the binop used for comparison. The visitor
only returns results for parameters that are supported by ReadySet.
Support is limited to expressions that directly compare a column and a
placeholder variable, rather than comparisons between general
expressions and variables.

The visitor is used to traverse the WHERE clause of SELECT, UPDATE, and
DELETE statements. Parameters in other parts of the query are not
supported.

Change-Id: I90668fbee67cbbe4c3cd5bbaa804bfe2edf8b47c
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3673
Reviewed-by: Griffin Smith <griffin@readyset.io>
Tested-by: Buildkite CI