dataflow-expression: Change `||` handling based on dialect
The double pipe operator has completely different behavior under default
MySQL as compared to Postgres. This change checks the dialect during
lowering to either return JsonConcat under Postgres or an error under
MySQL (since `||` isn't implemented under MySQL yet).

Note that at some point we will likely want to implement the text
version of the `||` concatenation operator; at that point we'll need to
consider the surrounding types to infer the correct Dataflow operator to
use for `||`, but for now since we are only working on implementing JSON
concatenation we can put off those kinds of changes and just assume we
want the JsonConcat operator whenever concatenation is the correct
interpretation of `||`.

Change-Id: Ia646037f47c88b71ca704a418d4866262e147b4f
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3681
Tested-by: Buildkite CI
Reviewed-by: Nikolai Vazquez <nikolai@readyset.io>