Remove doc comments from clap sub-structs
When structs that are intended for embedding into a parent struct using
clap flatten have doc comments, those comments unfortunately get
displayed in the binary --help text.

This is confusing for users trying to read this help text. This commit
removes all doc comments from sub-structs intended for embedding in a
clap parent struct for the ReadySet adapter, and includes an explanation
in a comment on each explaining why there aren't doc comments on each
sub-struct.

Change-Id: Ib983e6528bdc3bc06ba9d0bd2721510bf9eea701
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3674
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>
Reviewed-by: Jesse Seldess <jesse@readyset.io>